+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+<a name="1.6.0"></a>
+## 1.6.0 (2018-05-01)
+
+* Fix: functional tests (#185) ([664bfe3](https://github.com/box/box-annotations/commit/664bfe3)), closes [#185](https://github.com/box/box-annotations/issues/185)
+* Fix: Functional tests + cleanup annotations after Features (#186) ([ea868c4](https://github.com/box/box-annotations/commit/ea868c4)), closes [#186](https://github.com/box/box-annotations/issues/186)
+* Fix: Use BRUI CSS for create dialog textarea (#184) ([8faef28](https://github.com/box/box-annotations/commit/8faef28)), closes [#184](https://github.com/box/box-annotations/issues/184)
+* New: Add basic UI tests for point & highlight comment annotations (#182) ([44597de](https://github.com/box/box-annotations/commit/44597de)), closes [#182](https://github.com/box/box-annotations/issues/182)
+
+
+
+<a name="1.5.0"></a>
+## 1.5.0 (2018-04-24)
+
+* Update: dependencies (#177) ([5599ca6](https://github.com/box/box-annotations/commit/5599ca6)), closes [#177](https://github.com/box/box-annotations/issues/177)
+* Chore: Add tests to select/delete drawing (#180) ([90af909](https://github.com/box/box-annotations/commit/90af909)), closes [#180](https://github.com/box/box-annotations/issues/180)
+* Chore: Ensure all new drawings get deleted in functional tests (#181) ([8339dbf](https://github.com/box/box-annotations/commit/8339dbf)), closes [#181](https://github.com/box/box-annotations/issues/181)
+* Chore: Prefix box-react-ui css with .ba-annotation-dialog (#178) ([29cc0fd](https://github.com/box/box-annotations/commit/29cc0fd)), closes [#178](https://github.com/box/box-annotations/issues/178)
+* Fix: Replace {1} with user name on drawing save (#179) ([408b495](https://github.com/box/box-annotations/commit/408b495)), closes [#179](https://github.com/box/box-annotations/issues/179)
+
+
+
 <a name="1.4.0"></a>
 ## 1.4.0 (2018-04-17)